requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openpyxl==3.1.5
pillow==11.2.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
ttkbootstrap==1.12.0
typing_extensions==4.13.2
Werkzeug==3.1.3
